# contract/v1/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Standalone accounts v1",
		"version": "0.3.5"
	},
	"components": {
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"paths": {
...
		"/api/accounts/v1/account": {
			"get": {
				"operationId": "account",
				"security": [
					{
						"bearer": []
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id": {
											"type": "string",
											"format": "uuid",
											"pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
										},
										"supportID": {
											"type": "string",
											"pattern": "^[A-Z]{2}-[0-9A-Z]{4}-[0-9A-Z]{4}$"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time",
											"pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
										},
										"access": {
											"type": "object",
											"properties": {
												"status": {
													"type": "string",
													"enum": [
														"active",
														"grace",
														"granted",
														"expired",
														"revoked",
														"suspended",
														"none"
													]
												},
												"active": {
													"type": "boolean"
												},
												"plan": {
													"type": "string"
												},
												"source": {
													"type": "string",
													"enum": [
														"subscription",
														"purchase",
														"grant"
													]
												},
												"environment": {
													"type": "string",
													"enum": [
														"Production",
														"Sandbox",
														"Xcode"
													]
												},
												"expiresAt": {
													"type": "string",
													"format": "date-time",
													"pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
												},
												"willRenew": {
													"type": "boolean"
												}
											},
											"required": [
												"status",
												"active"
											],
											"additionalProperties": false
										},
										"identities": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"kind": {
														"type": "string",
														"minLength": 1
													},
													"label": {
														"type": "string"
													}
												},
												"required": [
													"kind"
												],
												"additionalProperties": false
											}
										}
									},
									"required": [
										"id",
										"supportID",
										"createdAt",
										"access"
									],
									"additionalProperties": false
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"enum": [
												"invalid_request",
												"unauthorized",
												"account_suspended",
												"purchase_in_use",
												"identity_in_use",
												"transaction_invalid",
												"transaction_revoked",
												"code_not_found",
												"code_expired",
												"rate_limited",
												"verification_unavailable",
												"invalid_credentials",
												"email_in_use",
												"signin_unavailable",
												"password_too_short"
											]
										},
										"message": {
											"type": "string"
										},
										"minLength": {
											"type": "integer",
											"minimum": -9007199254740991,
											"maximum": 9007199254740991
										}
									},
									"required": [
										"error"
									],
									"additionalProperties": false
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "deleteAccount",
				"security": [
					{
						"bearer": []
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"ok": {
											"type": "boolean",
											"enum": [
												true
											]
										}
									},
									"required": [
										"ok"
									],
									"additionalProperties": false
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"enum": [
												"invalid_request",
												"unauthorized",
												"account_suspended",
												"purchase_in_use",
												"identity_in_use",
												"transaction_invalid",
												"transaction_revoked",
												"code_not_found",
												"code_expired",
												"rate_limited",
												"verification_unavailable",
												"invalid_credentials",
												"email_in_use",
												"signin_unavailable",
												"password_too_short"
											]
										},
										"message": {
											"type": "string"
										},
										"minLength": {
											"type": "integer",
											"minimum": -9007199254740991,
											"maximum": 9007199254740991
										}
									},
									"required": [
										"error"
									],
									"additionalProperties": false
								}
							}
						}
					}
				}
			}
		},
...
		"/api/accounts/v1/account/password": {
			"post": {
				"operationId": "setPassword",
				"security": [
					{
						"bearer": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"maxLength": 320,
										"format": "email",
										"pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
									},
									"password": {
										"type": "string",
										"minLength": 1,
										"maxLength": 200,
										"description": "A new password (register, reset, change) must be at least the server’s minimum (default 8), or 400 `password_too_short` with `minLength`; signing in never checks length."
									},
									"currentPassword": {
										"type": "string",
										"maxLength": 200
									}
								},
								"required": [
									"email",
									"password"
								],
								"additionalProperties": false
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"ok": {
											"type": "boolean",
											"enum": [
												true
											]
										}
									},
									"required": [
										"ok"
									],
									"additionalProperties": false
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"enum": [
												"invalid_request",
												"unauthorized",
												"account_suspended",
												"purchase_in_use",
												"identity_in_use",
												"transaction_invalid",
												"transaction_revoked",
												"code_not_found",
												"code_expired",
												"rate_limited",
												"verification_unavailable",
												"invalid_credentials",
												"email_in_use",
												"signin_unavailable",
												"password_too_short"
											]
										},
										"message": {
											"type": "string"
										},
										"minLength": {
											"type": "integer",
											"minimum": -9007199254740991,
											"maximum": 9007199254740991
										}
									},
									"required": [
										"error"
									],
									"additionalProperties": false
								}
							}
						}
					}
				}
			}
		},
		"/api/accounts/v1/account/unlink": {
			"post": {
				"operationId": "unlinkIdentity",
				"security": [
					{
						"bearer": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"kind": {
										"type": "string",
										"minLength": 1,
										"maxLength": 50
									}
								},
								"required": [
									"kind"
								],
								"additionalProperties": false
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id": {
											"type": "string",
											"format": "uuid",
											"pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
										},
										"supportID": {
											"type": "string",
											"pattern": "^[A-Z]{2}-[0-9A-Z]{4}-[0-9A-Z]{4}$"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time",
											"pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
										},
										"access": {
											"type": "object",
											"properties": {
												"status": {
													"type": "string",
													"enum": [
														"active",
														"grace",
														"granted",
														"expired",
														"revoked",
														"suspended",
														"none"
													]
												},
												"active": {
													"type": "boolean"
												},
												"plan": {
													"type": "string"
												},
												"source": {
													"type": "string",
													"enum": [
														"subscription",
														"purchase",
														"grant"
													]
												},
												"environment": {
													"type": "string",
													"enum": [
														"Production",
														"Sandbox",
														"Xcode"
													]
												},
												"expiresAt": {
													"type": "string",
													"format": "date-time",
													"pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
												},
												"willRenew": {
													"type": "boolean"
												}
											},
											"required": [
												"status",
												"active"
											],
											"additionalProperties": false
										},
										"identities": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"kind": {
														"type": "string",
														"minLength": 1
													},
													"label": {
														"type": "string"
													}
												},
												"required": [
													"kind"
												],
												"additionalProperties": false
											}
										}
									},
									"required": [
										"id",
										"supportID",
										"createdAt",
										"access"
									],
									"additionalProperties": false
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"enum": [
												"invalid_request",
												"unauthorized",
												"account_suspended",
												"purchase_in_use",
												"identity_in_use",
												"transaction_invalid",
												"transaction_revoked",
												"code_not_found",
												"code_expired",
												"rate_limited",
												"verification_unavailable",
												"invalid_credentials",
												"email_in_use",
												"signin_unavailable",
												"password_too_short"
											]
										},
										"message": {
											"type": "string"
										},
										"minLength": {
											"type": "integer",
											"minimum": -9007199254740991,
											"maximum": 9007199254740991
										}
									},
									"required": [
										"error"
									],
									"additionalProperties": false
								}
							}
						}
					}
				}
			}
		}
	}
}